Add fixed version of callback
This commit adds improved/fixed version for callback and connection.
The old (and wip) are to be removed in the next commit.
Complete reimplementation for callback to make it valid
in recursive scenarios. Basic unit testing success, no leaks.
To be tested more intensive. The unit tests are also rather
random atm.