Remove Flags conversion and bool operators
This design decision was done because those operators
were really error-prone and easy to misuse.
E.g. when the underalying type of the enum was int
and a flags objects was casted to unsigned int, it
would use the bool conversion operator which is
obviously NOT was would be expected.

Use flags.value() instead and cast it as needed.
It returns the underlying raw value.