Discard ultra-generic vec/mat operation concept
This replaces the generic vec operation functions with the
previous type-bound ones. There were too many problems/issues/
drawbacks and too little advantages to make this work.

The change is still not complete, a section of nytl/matOps
still has to be reworked but this already removes a lot
of clutter like nytl_assure and nytl now compiles (and
the unit tests works) on clang 5 (maybe 4 as well?).

On the downside, this removes the ability (and therefore the
headers) of creating dynamic vec/mat types, i.e. nytl/dynVec
is deleted.